// src/guardian/interfaces/guardian.interface.ts
export interface GuardianArticle {
  id: string;
  type: string;
  sectionId: string;
  sectionName: string;
  webPublicationDate: string;
  webTitle: string;
  webUrl: string;
  apiUrl: string;
  isHosted: boolean;
  pillarId: string;
  pillarName: string;
}

export interface GuardianApiResponse {
  response: {
    status: string;
    userTier: string;
    total: number;
    startIndex: number;
    pageSize: number;
    currentPage: number;
    pages: number;
    orderBy: string;
    results: GuardianArticle[];
  };
}

export interface GuardianContent {
  id: string;
  type: string;
  sectionId: string;
  sectionName: string;
  webPublicationDate: string;
  webTitle: string;
  webUrl: string;
  apiUrl: string;
  fields: {
    body: string;
  };
}

export interface GuardianApiByIdResponse {
  response: {
    status: string;
    userTier: string;
    total: number;
    content: GuardianContent;
    isHosted: boolean;
    pillarId: string;
    pillarName: string;
  };
}